feat(pointers): Add PointerInputSource for macOS
