+++ src/Uno.UWP/UI/Core/CoreWindow.macOS.cs
@@ -1,23 +1,12 @@
 #if __MACOS__
 using System;
-using System.Runtime.InteropServices;
-using System.Threading.Tasks;
 using AppKit;
-
-using Uno.Extensions;
-using Windows.Devices.Input;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.Foundation.Metadata;
-using Uno.Foundation.Extensibility;
-using Uno.Foundation.Logging;
 
 namespace Windows.UI.Core
 {
 	partial class CoreWindow
 	{
 		private readonly NSWindow _window;
-		private ICoreWindowExtension _coreWindowExtension = default!; // Init in partial ctor.
 
 		public CoreWindow(NSWindow window)
 			: this()
@@ -25,95 +14,10 @@ namespace Windows.UI.Core
 			_window = window;
 		}
 
-		partial void InitializePartial()
-		{
-			_coreWindowExtension = ApiExtensibility.CreateInstance<ICoreWindowExtension>(this);
-		}
-
 		/// <summary>
 		/// Gets a reference to the native macOS Window behind the <see cref="CoreWindow"/> abstraction.
 		/// </summary>
 		internal NSWindow NativeWindow => _window;
-
-		internal void RefreshCursor()
-		{
-			if (_coreWindowExtension is { })
-			{
-				_coreWindowExtension.PointerCursor = _coreWindowExtension.PointerCursor;
-			}
-		}
-	}
-
-
-	internal partial class CoreWindowExtension : ICoreWindowExtension
-	{
-		private bool _cursorHidden;
-		private CoreCursor _pointerCursor = new CoreCursor(CoreCursorType.Arrow, 0);
-
-		/// <inheritdoc />
-		public CoreCursor PointerCursor
-		{
-			get => _pointerCursor;
-			set
-			{
-				_pointerCursor = value;
-				RefreshCursor();
-			}
-		}
-
-		/// <inheritdoc />
-		public void ReleasePointerCapture(PointerIdentifier pointer) { }
-
-		/// <inheritdoc />
-		public void SetPointerCapture(PointerIdentifier pointer) { }
-
-		private void RefreshCursor()
-		{
-			if (PointerCursor == null)
-			{
-				if (!_cursorHidden)
-				{
-					NSCursor.Hide();
-					_cursorHidden = true;
-				}
-			}
-			else
-			{
-				if (_cursorHidden)
-				{
-					NSCursor.Unhide();
-					_cursorHidden = false;
-				}
-				switch (_pointerCursor.Type)
-				{
-					case CoreCursorType.Arrow:
-						NSCursor.ArrowCursor.Set();
-						break;
-					case CoreCursorType.Cross:
-						NSCursor.CrosshairCursor.Set();
-						break;
-					case CoreCursorType.Hand:
-						NSCursor.PointingHandCursor.Set();
-						break;
-					case CoreCursorType.IBeam:
-						NSCursor.IBeamCursor.Set();
-						break;
-					case CoreCursorType.SizeNorthSouth:
-						NSCursor.ResizeUpDownCursor.Set();
-						break;
-					case CoreCursorType.SizeWestEast:
-						NSCursor.ResizeLeftRightCursor.Set();
-						break;
-					default:
-						if (this.Log().IsEnabled(LogLevel.Warning))
-						{
-							this.Log().LogWarning($"Cursor type '{_pointerCursor.Type}' is not supported on macOS. Default cursor is used instead.");
-						}
-						NSCursor.ArrowCursor.Set();
-						break;
-				}
-			}
-		}
 	}
 }
 #endif
